Getting closer to something actually usable

+++ Usings/ShaderUsings.cs
@@ -1,4 +1,4 @@
 ﻿global using static Source.ShaderDefines;
 global using StateSnapshot_t = short;
 global using ShadowStateId_t = short;
-global using ShaderAPITextureHandle_t = nint;
+global using ShaderAPITextureHandle_t = uint;
